Split out logs and up command into separate files
